requirements.txt
absl-py==0.11.0
bert-pytorch==0.0.1a4
bioc==1.1.dev3
bllipparser==2016.9.11
boto3==1.16.11
botocore==1.19.11
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
configparser==5.0.1
cryptography==2.8
cycler==0.11.0
decorator==4.4.2
deprecation==2.1.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.13.1
einops==0.3.0
filelock==3.0.12
fsspec==0.8.4
fuzzywuzzy==0.18.0
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
jmespath==0.10.0
joblib==0.17.0
JPype1==0.6.3
kiwisolver==1.3.1
lxml==3.7.3
Markdown==3.3.3
matplotlib==3.3.2
mkl-fft==1.0.6
mkl-random==1.0.1
networkx==2.0
nltk==3.6.7
numpy>=1.23.0
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.1.4
pathlib2==2.3.7.post1
pathtools==0.1.2
Pillow==8.0.1
ply==3.11
promise==2.3
protobuf==3.13.0
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.1.0
pyparsing==3.0.7
PySocks==1.7.1
PyStanfordDependencies==0.3.1
python-dateutil==2.8.1
pytorch-lightning==1.0.5
pytorch-pretrained-bert==0.6.2
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.91
sentry-sdk==0.19.1
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tokenizers==0.9.3
torchxrayvision==0.0.18
tqdm==4.51.0
transformers==3.5.1
typing-extensions==3.7.4.3
urllib3
wandb==0.10.8
watchdog==0.10.3
Werkzeug==1.0.1
zipp==3.4.0
